Extract Base functionality from S3 Loading Strategy
